test tests & transition from non-GUI to only GUI-based tests
