+++ tests/test_pytest_main.py
@@ -24,48 +24,7 @@ if not os.path.exists(fetch_dir):
 proj_directory = fetch_dir + "/Example_IMC"
 np.random.default_rng(42)
 
-#def test_fetch_IMC():
-#    fetch_IMC_example(fetch_dir)
-
-def test_raw_to_img():
-    global image_proc
-    image_proc = ImageAnalysis(proj_directory, from_mcds = True)
-    image_proc.directory_object.makedirs()
-    #image_proc.raw_to_img(0.85)
-    images = [f"{proj_directory}/images/img/{i}" for i in sorted(os.listdir(proj_directory + "/images/img"))]
-    assert(len(images) == 10), "Wrong number of images exported to images/img"               ## all the images are transferred
-    shutil.rmtree(proj_directory + "/raw") ## don't need raw anymore
-
-def test_regionprops_write():
-    image_proc.directory_object.make_analysis_dirs("test_analysis")
-    input_img_folder = proj_directory + "/images/img"
-    input_mask_folder = proj_directory + "/masks/example_deepcell_masks"    # "/masks/instanseg_masks" 
-    image_proc.make_segmentation_measurements(input_img_folder = input_img_folder, input_mask_folder = input_mask_folder)
-    analysis_dir = image_proc.directory_object.Analyses_dir + "/test_analysis"
-    intensities_dir = analysis_dir + "/intensities"
-    assert(len(os.listdir(analysis_dir + "/regionprops")) == 10), "Wrong number of regionprops csv exported (expecting 10 to match the number of images)"
-    assert(len(pd.read_csv(intensities_dir + "/CRC_1_ROI_001.ome.csv") == 2177)), "Unexpected number of cells in image 1"
-
-def test_setup_analysis():
-    panel_file, metadata, Analysis_panel_dir, metadata_dir = image_proc.to_analysis()
-    panel_file.to_csv(Analysis_panel_dir)
-    metadata.to_csv(metadata_dir)
-    assert(os.listdir(image_proc.directory_object.Analysis_internal_dir + "/Analysis_fcs")[0].rfind(".fcs") != -1), "FCS files not in /main/Analysis_fcs!"
-    assert(len(metadata) == 10), "Automatically generated Metadata file's length does not match the number of FCS files in the experiment!"
-    assert("marker_class" in panel_file.columns), "Automatically generated Analysis_panel file should have a 'marker_class' column"
-    assert("Analysis_panel.csv" in os.listdir(image_proc.directory_object.Analysis_internal_dir)), "Analysis_panel.csv not written to the proper place!"
-    assert("metadata.csv" in os.listdir(image_proc.directory_object.Analysis_internal_dir)), "metadata.csv not written to the proper place!"
-    assert("condition" in list(pd.read_csv(image_proc.directory_object.Analysis_internal_dir + "/metadata.csv").columns)), "Automatically generated metadata.csv file must have a 'condition' column!"
-
 ########### CRITICAL! -- depends on test_img_proc having been run first!
-def test_setup_directories():
-    global Analysis_panel
-    Analysis_panel = proj_directory + "/Analyses/Analysis_panel.csv"
-    global metadata
-    metadata = proj_directory + "/Analyses/metadata.csv"
-    shutil.copyfile(Analysis_panel, proj_directory + "/Analyses/test_analysis/main/Analysis_panel.csv")
-    shutil.copyfile(metadata, proj_directory + "/Analyses/test_analysis/main/metadata.csv")
-
 def test_load():
     global my_analysis
     my_analysis = Analysis()
